% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{augmentDPCdemo}
\alias{augmentDPCdemo}
\title{Augment Italy DPC data with demographics, and also removes Sardinia provinces}
\usage{
augmentDPCdemo(dpc, demodata)
}
\arguments{
\item{dpc}{Italian per province timeseries data.}

\item{demodata}{Dataframe sourced from Eurostat that contains GDP per Capita and Density information on Italian regions.}
}
\value{
Augmented dataframe that contains both COVID-19 cases timeseries and eurostats.
}
\description{
Augment Italy DPC data with demographics, and also removes Sardinia provinces
}
