% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{augmentDPCweather}
\alias{augmentDPCweather}
\title{Augment a table with date, lat, and long columns with weather data collected
closest to the given time. \strong{Warning}: This function will be very slow,
since it queries the NOAA database often. Also, it may hang indefinitely due
to a presumed bug in the worldmet package. Multiple attempts may be
necessary. Use the \code{robust} utility function to automate the re-running of
this function.}
\usage{
augmentDPCweather(dpc)
}
\arguments{
\item{dpc}{Dataframe of DPC per province time series case data. Required columns: "date", "lat", "long".}
}
\value{
Transformed dataframe with new columns added: "air_temp", "dewpoint", "RH".
}
\description{
Augment a table with date, lat, and long columns with weather data collected
closest to the given time. \strong{Warning}: This function will be very slow,
since it queries the NOAA database often. Also, it may hang indefinitely due
to a presumed bug in the worldmet package. Multiple attempts may be
necessary. Use the \code{robust} utility function to automate the re-running of
this function.
}
