% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{augmentsingleweather}
\alias{augmentsingleweather}
\title{Augment a single group of data points with weather, using a list of station codes.
Query each station code for air temperature, dewpoint, and relative humidity, taking the closest result that is not NA.}
\usage{
augmentsingleweather(group, station)
}
\arguments{
\item{group}{A dataframe groupby object on a particular province, which we are pulling weather data for.}

\item{station}{A comma separated string of station codes, which are used by importNOAA to query for weather data from those stations.}
}
\value{
Augmented dataframe groupby object which now contains "air_temp", "RH", "dewpoint" as columns.
}
\description{
Augment a single group of data points with weather, using a list of station codes.
Query each station code for air temperature, dewpoint, and relative humidity, taking the closest result that is not NA.
}
