% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateForcingModel.R
\docType{data}
\name{cf_mod}
\alias{cf_mod}
\title{Compiled STAN model of the seasonal forcing model described in accompanying
report.}
\format{
An object of class \code{stanmodel} of length 1.
}
\usage{
cf_mod
}
\description{
Compiled STAN model of the seasonal forcing model described in accompanying
report.
}
\keyword{datasets}
