% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{closestweatherstations}
\alias{closestweatherstations}
\title{Get the closest weather stations to a latitude longitude point.}
\usage{
closestweatherstations(lat, long, n)
}
\arguments{
\item{lat}{Latitude}

\item{long}{Longitude}

\item{n}{Number of weather stations to get.}
}
\value{
The codes of the closest weather stations in a list format.
}
\description{
Get the closest weather stations to a latitude longitude point.
}
