% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{closestweatherstationsM}
\alias{closestweatherstationsM}
\title{Memoized equivalent of functions, such that repeated function calls are efficient.}
\usage{
closestweatherstationsM(lat, long, n)
}
\arguments{
\item{lat}{Latitude}

\item{long}{Longitude}

\item{n}{Number of weather stations to get.}
}
\value{
The codes of the closest weather stations in a list format.
}
\description{
Memoized equivalent of functions, such that repeated function calls are efficient.
}
