% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{collectData}
\alias{collectData}
\title{Augments the Italian per province data with GDP, weather, density data.
Produces a file data/dpc-augmented.csv which contains the required data for
this analysis.}
\usage{
collectData(rewriteall = FALSE)
}
\arguments{
\item{rewriteall}{Attempt to refresh all files even if already exist. If
TRUE, this function may need to be called multiple times due to \code{worldmet}
issues.}
}
\description{
Augments the Italian per province data with GDP, weather, density data.
Produces a file data/dpc-augmented.csv which contains the required data for
this analysis.
}
