% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{ecdc-COVID-19-up-to-date}
\alias{ecdc-COVID-19-up-to-date}
\title{ecdc-COVID-19-up-to-date}
\description{
Raw data for various countries published by the ECDC. Not used in the
analysis by default.
}
