% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{fetch_file}
\alias{fetch_file}
\title{Fetches a raw file from url, and saves it to a filename.
For proper style, filename string should begin with "data/".
Throws exception on failure.}
\usage{
fetch_file(url, filename)
}
\arguments{
\item{url}{The URL of the file to download.}

\item{filename}{The filename that the file is saved to.}
}
\value{
None
}
\description{
Fetches a raw file from url, and saves it to a filename.
For proper style, filename string should begin with "data/".
Throws exception on failure.
}
