% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{fetch_latest_csse}
\alias{fetch_latest_csse}
\title{Fetches latest data from John Hopkins University CSSE.}
\usage{
fetch_latest_csse()
}
\value{
None
}
\description{
Fetches latest data from John Hopkins University CSSE.
}
