% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{fetch_latest_dpc}
\alias{fetch_latest_dpc}
\title{Fetch Italian provinces data from DPC.}
\usage{
fetch_latest_dpc()
}
\value{
None

Rename Italian column names
}
\description{
Fetch Italian provinces data from DPC.
}
