% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{fetch_latest_ecdc}
\alias{fetch_latest_ecdc}
\title{Fetches the latest COVID-19 cases data from European Center for Disease Control (ECDC).}
\usage{
fetch_latest_ecdc()
}
\value{
None

Quick cleanup on the column titles.
}
\description{
Fetches the latest COVID-19 cases data from European Center for Disease Control (ECDC).
}
