% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{getdemodata}
\alias{getdemodata}
\title{Fetch GDP per capita data and population density data for Italian provinces
Source: Eurostat}
\usage{
getdemodata()
}
\value{
Eurostat dataframe with columns "name", "gdppercapita", "density", "population".
Get GDP/capita data from Eurostat.

Get population density data from Eurostat.
Get population data from Eurostat.
}
\description{
Fetch GDP per capita data and population density data for Italian provinces
Source: Eurostat
}
