% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{getprovincelist}
\alias{getprovincelist}
\title{Obtain mapping of health ministry province names to eurostat codes}
\usage{
getprovincelist()
}
\value{
Dataframe of italian provinces with columns "name", "code".
}
\description{
Obtain mapping of health ministry province names to eurostat codes
}
