% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateForcingModel.R
\name{getSIdist}
\alias{getSIdist}
\title{Discretize the serial interval distribution of 4.7 std. dev 2.9, used in the
CMMID paper https://epiforecasts.io/covid/methods.html}
\usage{
getSIdist(n, mean = 4.7, sd = 2.9)
}
\arguments{
\item{n}{The number of discrete dates in the future to compute a serial
interval density for.}

\item{mean}{The estimated mean of the serial interval.}

\item{sd}{The estimated standard distribution of the serial interval.}
}
\value{
A vector whose index corresponds to the number of days in the future,
and whose value corresponds to the density of the serial interval
distribution.
}
\description{
Discretize the serial interval distribution of 4.7 std. dev 2.9, used in the
CMMID paper https://epiforecasts.io/covid/methods.html
}
