% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{importNOAAM}
\alias{importNOAAM}
\title{ImportNOAA is a function exposed in "worldmet" library. Gets data from weather stations.}
\usage{
importNOAAM(
  code = "037720-99999",
  year = 2014,
  hourly = TRUE,
  precip = FALSE,
  PWC = FALSE,
  parallel = TRUE,
  quiet = FALSE,
  path = NA
)
}
\description{
ImportNOAA is a function exposed in "worldmet" library. Gets data from weather stations.
}
