% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{mean_cl_quantile}
\alias{mean_cl_quantile}
\title{Given a series, return the mean and two specified quantiles.
Code copied from https://stackoverflow.com/a/51993331/3575047 written by @www.}
\usage{
mean_cl_quantile(x, q = c(0.1, 0.9), na.rm = TRUE)
}
\arguments{
\item{x}{The series to process.}

\item{q}{The limits of confidence (e.g. 0.1 to 0.9 for 80\%)}

\item{na.rm}{Whether to ignore NA values.}
}
\value{
A gg-plot geom_smooth-usable summary of the mean and quantiles.
}
\description{
Given a series, return the mean and two specified quantiles.
Code copied from https://stackoverflow.com/a/51993331/3575047 written by @www.
}
