% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_beta_confidence}
\alias{plot_beta_confidence}
\title{Given the Bayesian posterior stanfit object, plot the MCMC confidence for
beta.}
\usage{
plot_beta_confidence(res)
}
\arguments{
\item{res}{Output of \code{results()} containing results of the analysis.}
}
\value{
A plot of location of standardized beta parameters, with 50\%
confidence interval highlighted.
}
\description{
Given the Bayesian posterior stanfit object, plot the MCMC confidence for
beta.
}
