% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_epsilon_dispersion}
\alias{plot_epsilon_dispersion}
\title{Plot a histogram of all samples of epsilon across all times and locations.}
\usage{
plot_epsilon_dispersion(res)
}
\arguments{
\item{res}{Samples of posterior distribution of parameters.}
}
\value{
A histogram of all samples of epsilon.
}
\description{
Plot a histogram of all samples of epsilon across all times and locations.
}
