% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_independentR}
\alias{plot_independentR}
\title{Given the Bayesian posterior samples and initial data, return the modeled
time-varying R independent of climate forcings for the entire country.}
\usage{
plot_independentR(res)
}
\arguments{
\item{res}{Samples of posterior distribution of parameters.}
}
\value{
A ggplot2 time series of estimated independent R with 80\% confidence.
}
\description{
Given the Bayesian posterior samples and initial data, return the modeled
time-varying R independent of climate forcings for the entire country.
}
