% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_locationR}
\alias{plot_locationR}
\title{Given the Bayesian posterior samples and initial data, return the modeled
time-varying R for location with the given index.}
\usage{
plot_locationR(params, loc)
}
\arguments{
\item{loc}{Index of the location to examine.}

\item{res}{Samples of posterior distribution of parameters.}
}
\value{
A ggplot2 time series of estimated R with 80\% confidence.
}
\description{
Given the Bayesian posterior samples and initial data, return the modeled
time-varying R for location with the given index.
}
