% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{removeyear}
\alias{removeyear}
\title{Create labels for dates appropriate for use in a chart, by removing the year.}
\usage{
removeyear(range)
}
\arguments{
\item{range}{The date range.}
}
\value{
A vector of strings to use as labels for a chart.
}
\description{
Create labels for dates appropriate for use in a chart, by removing the year.
}
