% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisDetails.R
\name{results}
\alias{results}
\title{Return a list containing results of the analysis, sufficient for generating plots.
Must be called after \code{collectData()} and \code{runMCMC()} and \code{runThompson()}.}
\usage{
results()
}
\value{
The saved results of the analysis.
}
\description{
Return a list containing results of the analysis, sufficient for generating plots.
Must be called after \code{collectData()} and \code{runMCMC()} and \code{runThompson()}.
}
