% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateForcingModel.R
\name{runMCMC}
\alias{runMCMC}
\title{Compile the stan model, run MCMC, and save results.}
\usage{
runMCMC(cutoff = "2020-03-14")
}
\arguments{
\item{cutoff}{The cutoff date (default: Mar 14).}
}
\description{
Compile the stan model, run MCMC, and save results.
}
