% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{time_series_covid19_recovered_global}
\alias{time_series_covid19_recovered_global}
\title{time_series_covid19_recovered_global}
\description{
Raw data for recovery counts from various countries published by the Johns
Hopkins school. Not used in the analysis by default.
}
