% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{transform_total_cases}
\alias{transform_total_cases}
\title{Repair total case data to be monotonically increasing, through taking a rolling maximum.
Drop the last 4 dates as it is likely to have no weather data for each location.
Add new_cases column which contains the number of new cases each day.}
\usage{
transform_total_cases(dpc)
}
\arguments{
\item{dpc}{Dataframe of DPC per province timeseries. Requires the columns: "date", "province", "total_cases".}
}
\value{
Transformed dataframe with modified "total_cases" column, where the column is now monotonically increasing.
}
\description{
Repair total case data to be monotonically increasing, through taking a rolling maximum.
Drop the last 4 dates as it is likely to have no weather data for each location.
Add new_cases column which contains the number of new cases each day.
}
