% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPipeline.R
\name{closestweatherstationsM}
\alias{closestweatherstationsM}
\title{Memoized equivalent of functions, such that repeated function calls are efficient.}
\usage{
closestweatherstationsM(lat, long, n)
}
\description{
Memoized equivalent of functions, such that repeated function calls are efficient.
}
